+++ app/upgrades/v8rc0/constants.go
@@ -1,20 +1,21 @@
-package v7rc0
+package v8rc0
 
 import (
 	"cosmossdk.io/store/types"
-	"github.com/MANTRA-Chain/mantrachain/v7/app/upgrades"
+	"github.com/MANTRA-Chain/mantrachain/v8/app/upgrades"
+	providertypes "github.com/cosmos/interchain-security/v7/x/ccv/provider/types"
 )
 
 const (
 	// UpgradeName defines the on-chain upgrade name.
-	UpgradeName = "v7.0.0-rc0"
+	UpgradeName = "v8.0.0-rc0"
 )
 
 var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,
 	StoreUpgrades: types.StoreUpgrades{
-		Added:   []string{},
+		Added:   []string{providertypes.ModuleName},
 		Deleted: []string{},
 	},
 }