+++ cmd/mantrachaind/cmd/commands.go
@@ -10,7 +10,7 @@ import (
 	"github.com/CosmWasm/wasmd/x/wasm"
 	wasmcli "github.com/CosmWasm/wasmd/x/wasm/client/cli"
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
-	"github.com/MANTRA-Chain/mantrachain/v7/app"
+	"github.com/MANTRA-Chain/mantrachain/v8/app"
 	cmtcli "github.com/cometbft/cometbft/libs/cli"
 	dbm "github.com/cosmos/cosmos-db"
 	"github.com/cosmos/cosmos-sdk/client"
@@ -69,6 +69,9 @@ func initRootCmd(
 		pruning.Cmd(sdkAppCreator, app.DefaultNodeHome),
 		confixcmd.ConfigCommand(),
 		snapshot.Cmd(sdkAppCreator),
+		ExportBlockEventsCmd(),
+		ImportBlockEventsCmd(),
+		CleanupBlockEventsCmd(),
 	)
 
 	// add server commands