+++ tests/e2e/e2e_sanction_test.go
@@ -25,7 +25,7 @@ func (s *IntegrationTestSuite) writeAddBlacklistAccountsProposal(c *chain, block
 		 }
 		],
 		"metadata": "ipfs://CID",
-		"deposit": "100uom",
+		"deposit": "100000000000000amantra",
 		"title": "Add %s to blacklist",
 		"summary": "e2e-test adding to blacklist"
 	   }`
@@ -94,7 +94,7 @@ func (s *IntegrationTestSuite) writeRemoveBlacklistAccountsProposal(c *chain, bl
 		 }
 		],
 		"metadata": "ipfs://CID",
-		"deposit": "100uom",
+		"deposit": "100000000000000amantra",
 		"title": "Remove %s to blacklist",
 		"summary": "e2e-test remove from blacklist"
 	   }`