+++ tests/e2e/genesis.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"cosmossdk.io/math"
-	tokenfactorytypes "github.com/MANTRA-Chain/mantrachain/v7/x/tokenfactory/types"
+	tokenfactorytypes "github.com/MANTRA-Chain/mantrachain/v8/x/tokenfactory/types"
 	"github.com/cosmos/cosmos-sdk/server"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
@@ -99,7 +99,7 @@ func modifyGenesis(path, moniker, amountStr string, addrAll []sdk.AccAddress, ba
 	appState[stakingtypes.ModuleName] = stakingGenStateBz
 
 	// Refactor to separate method
-	amnt := math.NewInt(10000)
+	amnt := math.NewInt(10000000000000000)
 	quorum, _ := math.LegacyNewDecFromStr("0.000000000000000001")
 	threshold, _ := math.LegacyNewDecFromStr("0.000000000000000001")