+++ tests/interchain/chainsuite/config.go
@@ -31,14 +31,14 @@ type SuiteConfig struct {
 
 const (
 	CommitTimeout          = 4 * time.Second
-	Uom                    = "uom"
-	GovMinDepositAmount    = 1000
-	GovDepositAmount       = "5000000" + Uom
+	Amantra                = "amantra"
+	GovMinDepositAmount    = 1000000000000000
+	GovDepositAmount       = "5000000000000000000" + Amantra
 	GovDepositPeriod       = 60 * time.Second
 	GovVotingPeriod        = 40 * time.Second
 	DowntimeJailDuration   = 10 * time.Second
 	ProviderSlashingWindow = 10
-	GasPrices              = "0.01" + Uom
+	GasPrices              = "10000000000" + Amantra
 	ValidatorCount         = 6
 	UpgradeDelta           = 12
 	ValidatorFunds         = 11_000_000_000
@@ -113,7 +113,7 @@ func DefaultSuiteConfig(env Environment) SuiteConfig {
 			NumValidators: &validators,
 			Version:       env.OldMantraImageVersion,
 			ChainConfig: ibc.ChainConfig{
-				Denom:         Uom,
+				Denom:         Amantra,
 				GasPrices:     GasPrices,
 				Gas:           "auto",
 				GasAdjustment: 2.0,
@@ -133,7 +133,7 @@ func DefaultSuiteConfig(env Environment) SuiteConfig {
 				TrustingPeriod:       "48h",
 				NoHostMount:          false,
 				ModifyGenesis:        cosmos.ModifyGenesis(DefaultGenesis()),
-				ModifyGenesisAmounts: DefaultGenesisAmounts(Uom),
+				ModifyGenesisAmounts: DefaultGenesisAmounts(Amantra),
 			},
 		},
 	}
@@ -153,7 +153,7 @@ func DefaultGenesis() []cosmos.GenesisKV {
 	return []cosmos.GenesisKV{
 		cosmos.NewGenesisKV("app_state.gov.params.voting_period", GovVotingPeriod.String()),
 		cosmos.NewGenesisKV("app_state.gov.params.max_deposit_period", GovDepositPeriod.String()),
-		cosmos.NewGenesisKV("app_state.gov.params.min_deposit.0.denom", Uom),
+		cosmos.NewGenesisKV("app_state.gov.params.min_deposit.0.denom", Amantra),
 		cosmos.NewGenesisKV("app_state.gov.params.min_deposit.0.amount", strconv.Itoa(GovMinDepositAmount)),
 		cosmos.NewGenesisKV("app_state.slashing.params.signed_blocks_window", strconv.Itoa(ProviderSlashingWindow)),
 		cosmos.NewGenesisKV("app_state.slashing.params.downtime_jail_duration", DowntimeJailDuration.String()),