+++ x/tax/module/genesis_test.go
@@ -3,11 +3,11 @@ package tax_test
 import (
 	"testing"
 
-	appparams "github.com/MANTRA-Chain/mantrachain/v7/app/params"
-	keepertest "github.com/MANTRA-Chain/mantrachain/v7/testutil/keeper"
-	"github.com/MANTRA-Chain/mantrachain/v7/testutil/nullify"
-	tax "github.com/MANTRA-Chain/mantrachain/v7/x/tax/module"
-	"github.com/MANTRA-Chain/mantrachain/v7/x/tax/types"
+	appparams "github.com/MANTRA-Chain/mantrachain/v8/app/params"
+	keepertest "github.com/MANTRA-Chain/mantrachain/v8/testutil/keeper"
+	"github.com/MANTRA-Chain/mantrachain/v8/testutil/nullify"
+	tax "github.com/MANTRA-Chain/mantrachain/v8/x/tax/module"
+	"github.com/MANTRA-Chain/mantrachain/v8/x/tax/types"
 	"github.com/stretchr/testify/require"
 )