+++ x/tokenfactory/keeper/grpc_query.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/MANTRA-Chain/mantrachain/v7/x/tokenfactory/types"
+	"github.com/MANTRA-Chain/mantrachain/v8/x/tokenfactory/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )